refactor: refactor arrow conversion
