+++ src/common/src/array/arrow/arrow_impl.rs
@@ -42,8 +42,8 @@
 
 use std::fmt::Write;
 
+use arrow_array::array;
 use arrow_array::cast::AsArray;
-use arrow_array_iceberg::array;
 use arrow_buffer::OffsetBuffer;
 use chrono::{DateTime, NaiveDateTime, NaiveTime};
 use itertools::Itertools;
