+++ src/connector/Cargo.toml
@@ -16,12 +16,6 @@ normal = ["workspace-hack"]
 [dependencies]
 anyhow = "1"
 apache-avro = { workspace = true }
-arrow-array = { workspace = true }
-arrow-array-iceberg = { workspace = true }
-arrow-row = { workspace = true }
-arrow-schema = { workspace = true }
-arrow-schema-iceberg = { workspace = true }
-arrow-select = { workspace = true }
 assert_matches = "1"
 async-compression = { version = "0.4.5", features = ["gzip", "tokio"] }
 async-nats = "0.35"
