+++ src/expr/core/src/table_function/user_defined.rs
@@ -15,7 +15,7 @@
 use std::sync::Arc;
 
 use anyhow::Context;
-use arrow_schema::{Fields, Schema, SchemaRef};
+use risingwave_common::array::arrow::arrow_schema_udf::{Fields, Schema, SchemaRef};
 use risingwave_common::array::arrow::{FromArrow, ToArrow, UdfArrowConvert};
 use risingwave_common::array::I32Array;
 use risingwave_common::bail;
