+++ src/expr/impl/Cargo.toml
@@ -24,10 +24,7 @@ wasm-udf = ["arrow-udf-wasm", "zstd"]
 [dependencies]
 aho-corasick = "1"
 anyhow = "1"
-arrow-array = { workspace = true }
 arrow-flight = { workspace = true, optional = true }
-arrow-schema = { workspace = true }
-arrow-schema-iceberg = { workspace = true }
 arrow-udf-flight = { workspace = true, optional = true }
 arrow-udf-js = { workspace = true, optional = true }
 arrow-udf-python = { workspace = true, optional = true }
